[KTNB] add a notification for missing JDK
This commit adds an editor notification that appears when JDK for running Kotlin Notebooks is not selected or is missing. Notifications are updated by the KotlinNotebookMissingJdkService, which listens for changes in the selected project jdk, project jdk table (which can be used as a fallback) and selected jdk in the KotlinNotebookProjectOptionsProvider to update these notifications.

Space-RevId: 38be5bb56257e35a0cd4c3f1bfd915255a5e37ee