DS-5356 Repair Jupyter Lab managed servers by reimplementing version stats collector
The bug was introduced in d3237315531f489d190b654d800bb81b32446986: the change at JupyterNotebookProcessHandler.kt:129 was incorrect and broke just announced Jupyter Lab managed servers support. That change had been done only for making JupyterGetServerVersionTest green.

After returning the old code back and extending JupyterGetServerVersionTest with Jupyter Lab, it was found that JupyterConnectionsCollector sends deceptive stats for Jupyter Lab. The call `/api` returns the version of the Jupyter Lab server instead of the Jupyter Notebook server. Obviously, those servers are different and they shouldn't be put into a single bin. This fact led to reorganizing the collector: there left no ways to obtain the server version without authentication. So, now stats is reported only after successful authentication to the server, in contrast with the previous approach when stats had been reported as soon as the existence of a server was discovered.

Also, since there was the need to change the API of `JupyterClient.getServerVersion` anyway, it was decided to improve version stats reporting for JupyterHub. Hub is not a server by itself, it's an orchestrator for other servers. Now the statistics collector reports both the version of JupyterHub and the version of the Notebook/Lab server launched by JupyterHub.

Space-RevId: cc5e4592bb4af96c744c54f92dce9be55bde3c27