KTNB-370 [KTNB Outputs] Handle empty dataframe in Jupyter Kotlin plugin
Added checks and handlers for edge case where dataframe being rendered in the Jupyter Kotlin plugin is empty. Previously, this caused errors and failed to render a usable output. Now, in the case of an empty dataframe, we create an empty table with the column headers present (if any), treating it as a valid case.

Space-RevId: 8d8a088805dd25100a95e49da012dde371fe7022