KTNB-383 [KTNB Outputs] Add error notification for non-comparable column sorting in Kotlin notebooks
A notification group "Kotlin Notebook output error" was added in the plugin.xml file. In KotlinDataframeTableDataProvider.kt, an exception handling was added for when a JsonParseException occurs during the parsing of a data frame output. Specifically, if the parsing fails due to columns that cannot be compared, a warning notification will be displayed, and a RuntimeException will be thrown. The error message for this case "Could not sort by column" was added in KotlinNotebookBundle.properties. This improves user experience by providing clear feedback when an operation isn't possible.

Space-RevId: fef98f8527333ea87aedeb5c55dea9bc557333de