KTNB-383 [KTNB Outputs] Update error handling and notification in Kotlin Jupyter plugin
The error notification and handling system of the Kotlin Jupyter plugin have been updated to provide more specific and helpful error messages for the user. The notification group key in the plugin.xml file was changed to more accurately reflect its purpose.

In the KotlinDataframeTableDataProvider.kt file, instead of throwing a generic RuntimeException, a new DataFrameParseException is thrown when there is an error in parsing DataFrame output. This exception will provide clearer information about the issue. Furthermore, an additional function was introduced to extract erroneous column names to enhance the clarity of error descriptions.

Also, the corresponding English language bundle file was updated to reflect these changes in error and notification messages.

Space-RevId: 6d89eb313ec2f0744dfe4d54a4bb29d8426be272