KTNB-382 [KTNB Outputs] Add compatibility for various DataFrame types when sorting by column in UI.
The update resolves the limitations of the existing code that could only process standard DataFrame types by expanding support for various intermediate object types that may occur during runtime in the Kotlin Notebooks plugin. This was achieved by attempting to convert object to Dataframe in a try-catch block, ensuring compatibility with DataFrame versions <= 0.12.1 and is a temporary fix for issue KTNB-382.

Space-RevId: 143703e85c6d5d2d137b258116045ddbe27f26d4