KTNB-693 Add support for the "Explain DataFrame" action in Kotlin Notebooks.
This MR adds support for the "Explain DataFrame" action similar to what is available in PyCharm and Jupyter Notebooks.

There are a few noticeable differences, though.

1. Experiments show that adding the output of describe() to the prompt makes the LLM answer much more useful. So this MR also adds support for DSDataDescription

2. I added a "Please add 3 suggestions for follow-up tasks" to the prompt. It seems to make the response much more actionable.

There are still issues with the current architecture as this functionality depends on the scientific.table module, which is almost 100% about Python support. It would probably be useful to make this module language agnostic, but it seemed out of scope for this MR.

Merge-request: IJ-MR-136005
Merged-by: Christian Melchior <christian.melchior@jetbrains.com>

Space-RevId: c6e9cd6d6f98d100e5de9fbdb81785931cd84896