KTNB: try to fix flaky timeouts
Skipped values in the `scriptsUpdateCompleted` flow could cause a deadlock during `setUpDependenciesSynchronously()`, when we emit the stopping value before the subscriber is registered, and then the subscriber waits for it infinitely.
`extraBufferCapacity` is only applicable when there are slow subscribers. Values emitted when there are no subscribers yet are only saved to the `replay` buffer.

Space-RevId: 6b3940d8d4cb561d946fac7bebd911ceac1fff9a