actions: fix in-pace replacements for overridden ActionToolbar button presentations
We do not expect to face buttons created by some other means in the 'isDefaultActionButtonImplementation'.
So if the 'createToolbarButton' is overridden unconditionally, 'return true' is a valid answer.
For clarity and future-robustness, the 'return oldActionButton.javaClass == MyActionButton::class.java' is used instead where possible.

This fixes '... toolbar creates new components for 20 updates in a row. The latest button is created for ...' warnings.

Space-RevId: ae7bdf8a46f93db2e61dbd56dbe6f917e9680ca2