[scientific-tables] Introduce DSTableLoadingInterruptedException
Add DSTableLoadingInterruptedException to distinguish table loading interruptions
from other data exceptions. This allows for different handling - interruptions are
logged at WARN level (e.g., when execution is aborted) while other errors are logged
at ERROR level.

Updated all table data providers to throw and handle the new exception type.

This change is done in the scope of the DS-7201. The reason for it is that the newly created env test for the notebooks (not yet committed) was flaky because of this behaviour:

- We run the SQL cell and generate the table data
- The output data is verified, and the test finishes and disposes
- However, when the output is updated, the listener is called that creates a UI component out of this data. During the UI component creation, we create a second request to the notebooks. However, sometimes we finish the test before this process is finished, and thus the request is aborted.

The abortion of the request is a normal process, however, in the `loadDynamicTableData` we rethrow the original interrupt exception. This exception is not mentioned in documentation or `@Throws` annotation and is not properly processed by the caller. In order to handle this issue, the new type of the exception is created. It's thrown when the request to the notebook is aborted or fails and it's processed properly by the caller.

Space-RevId: beda06220527a3a5ec243605355fe590e22cb626